+++ tests/unit/classes/preload/classes/CH.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace mako\tests\unit\classes\preload\classes;
+
+class CH
+{
+	protected string $foo;
+	protected CB $cb;
+}