+++ tests/unit/classes/preload/classes/CI.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace mako\tests\unit\classes\preload\classes;
+
+class CI
+{
+	protected IA|IB $iaib;
+}